Move Cache and SynchronizedCache to its own module in a new util package
