+++ clock/finder/zone_finder/provider.py
@@ -7,7 +7,7 @@ from clock.finder.zone_finder.zone_finders.country import CountryZoneFinder
 from clock.finder.zone_finder.zone_finders.localized import LocalizedZoneFinder
 from clock.finder.zone_finder.zone_finders.localized_date_time import LocalizedDateTimeZoneFinder
 from clock.finder.zone_finder.zone_finders.name import NameZoneFinder
-from clock.util import Cache
+from clock.util.cache import Cache
 
 
 class ZoneFindersProvider:
