+++ clock/util/cache.py
